this is the response of reccomendations
Okay, a casual party! Great. For a casual party, I'd recommend a stylish but relaxed outfit. How about a nice pair of dark-wash jeans or chinos, paired with a button-down shirt or a polo shirt? You could also add a casual blazer or a stylish jacket depending on the weather. What do you think about those ideas?

seeing this in summary cut offs the vibe tell him reply in points